Move slice functions to where they belong
